requirements.txt
torch>=1.7.0
torchvision>=0.8.0
Pillow>=8.0.0
numpy>=1.19.0
segmentation-models-pytorch>=0.2.0
albumentations>=1.0.0 